requirements.txt
streamlit==1.28.0
pypdf==3.17.0
chromadb==0.4.15
sentence-transformers==2.2.2
google-generativeai==0.3.0
python-dotenv==1.0.0
langchain-text-splitters==0.0.1
langchain-community==0.0.10